Refactor image funcs for seperate file
